ALWAYS initialise your struct fields folks. zerofication fixed
due to missing initialisation of the field `.coeff` in the case when
the Poly is not a coeff, some cases of zerofication did not work. in
particular you'd get as a result a poly with a NULL and a trash val as
its coeff instead of zero. also valgrind is now a little bit less
troublesome