if insufficiently many polys in q yet exp = 0 then return 1, not 0
as specified in the task when there's not enough q polys we should put
in 0 instead for x_i. BUT what if there's x_i^0 there where we put
this x_i=0? Well, then we do not simply prune this poly-tree branch
but rather check whether we get a 1 there from 0^0 or 0 from 0^n n!=0